Consolidate string and severity util functions
