+++ frontend/src/pages/VulnerabilityScanDash/TopVulnerabilities.tsx
@@ -7,7 +7,7 @@ import Typography from '@mui/material/Typography';
 import Circle from '@mui/icons-material/Circle';
 import { CVEItem } from 'types/vuln-scan-stats';
 import RoundedTable from 'components/Dashboard/RoundedTable';
-import { severityColor } from 'utils/severityLevelColorMap';
+import { severityColor } from 'utils/getSeverityColor';
 import TruncatedLink from 'components/Dashboard/TruncatedLink';
 import GraphChip from 'components/Dashboard/GraphChip';
 import InfoLabel from 'components/Dashboard/InfoLabel';
