+++ frontend/src/pages/VulnerabilityScanDash/TopVulnerableHosts.tsx
@@ -4,7 +4,7 @@ import { useHistory } from 'react-router-dom';
 import Box from '@mui/material/Box';
 import { useTheme } from '@mui/material/styles';
 import { axisClasses, BarChart, chartsGridClasses } from '@mui/x-charts';
-import { getSeverityLevelColorMap } from 'utils/severityLevelColorMap';
+import { getSeverityLevelColorMap } from 'utils/getSeverityColor';
 import GraphChip from 'components/Dashboard/GraphChip';
 import InfoLabel from 'components/Dashboard/InfoLabel';
 import infoIconContent from './infoIconContent.json';
