--- frontend/src/utils/matchPath.ts
@@ -1,4 +1,0 @@
-export const matchPath = (paths: string[], path: string) => {
-  if (paths.includes(path)) return true;
-  return false;
-};
